Build-KnowlegeGraph/pyproject.toml
[project]
name = "build-knowlegegraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=2.1.5",
    "langchain>=0.3.25",
    "langchain-experimental>=0.3.4",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.18",
    "neo4j>=5.28.1",
    "python-dotenv>=1.1.0",
    "langchain-ollama>=0.3.3",
    "langchain-community>=0.3.24",
    "pypdf>=5.5.0",
    "ipykernel>=6.29.5",
    "unstructured>=0.17.2",
    "pdfminer>=20191125",
]
